refactor: useAuth.tsx -> auth.ts
